mise en place de la boite de confirmation de suppresion de compte

+++ profil.php
@@ -12,23 +12,12 @@ $users = $sql->fetch(PDO::FETCH_ASSOC);
 $sql = $pdo->prepare("SELECT inscription.deck_name FROM inscription WHERE user_id = :id");
 $sql->execute(['id' => $_SESSION['id']]);
 $decks = $sql->fetchAll(PDO::FETCH_ASSOC);
-var_dump($users['id']);
+var_dump("./deleProfil.php?id=".$users['id']);
 
 
 $title = "Page de profil";
 require "./template/header.php";
 ?>
-<!-- mise en place d'une fonction contenant une alerte à confirmer si l'on veut supprimer son compte -->
-<script>
-    function confirmDelete(event) {
-        event.preventDefault();
-        var useConfirm = confirm("Etes vous sur de vouloir supprimer votre compte?")
-
-        if (useConfirm) {
-            window.location.href = event.target.href;
-        }
-    }
-</script>
 <div class="profil">
     <h2>Page de profil</h2>
     <!-- On affiche les données de la bdd avec htmlspecialchars pour éviter le XSS -->
@@ -46,9 +35,19 @@ require "./template/header.php";
     </div>
     <!-- lien vers la deconnexion et la suppresion du compte -->
     <a href="./deconnexion.php"><button>Deconnexion</button></a>
-    <a href="./deleteProfil.php?id=<?php echo $users["id"] ?>" ><button>Supprimer son compte</button></a>
+    <a href="./deleteProfil.php?id=<?php echo $users["id"]; ?>" onclick="confirmDelete(event)" ><button>Supprimer son compte</button></a>
 </div>
+<!-- mise en place d'une fonction contenant une alerte à confirmer si l'on veut supprimer son compte -->
+<script>
+    function confirmDelete(event) {
+        event.preventDefault();
+        var useConfirm = confirm("Etes vous sur de vouloir supprimer votre compte?")
 
+        if (useConfirm) {
+            window.location.href = event.currentTarget.href;
+        }
+    }
+</script>
 
 <?php
 require "./template/footer.php";
